+++ script/vlCVX/3_merkles/delegators.ts
@@ -1,0 +1,4 @@
+/*
+TODO:
+- Merkle with SDT + CRVusd ONLY for forwarded delegators
+*/
