+++ script/vlCVX/3_merkles/nonDelegators.ts
@@ -1,0 +1,4 @@
+/*
+TODO:
+- Merkle with regular tokens + SDT for delegators (non-forwarded)
+*/
